Issue #4437: Modified AvoidStarImportCheckTest.java and moved its input files to the avoidstarimport subdirectory
