+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/imports/avoidstarimport/InputAvoidStarImportDefault.java
@@ -2,7 +2,7 @@
 // Test case file for checkstyle.
 // Created: 2001
 ////////////////////////////////////////////////////////////////////////////////
-package com.puppycrawl.tools.checkstyle.checks.imports;
+package com.puppycrawl.tools.checkstyle.checks.imports.avoidstarimport;
 
 import com.puppycrawl.tools.checkstyle.checks.imports.*;
 
@@ -53,7 +53,7 @@ import com.puppycrawl.tools.checkstyle.DefaultLogger;
  * @author Michael Studman
  * @see Calendar Should avoid unused import for Calendar
  **/
-class InputAvoidStarImport
+class InputAvoidStarImportDefault
 {
     /** ignore **/
     private Class mUse1 = Connection.class;
